refactor: move `Ease` to `libAnki` module
Removes a cyclic dependency: libAnki <-> AnkiDroid

Better than moving to CardAnswer.Rating due to
CardAnswer.Rating.UNRECOGNIZED

Issue 18015
