+++ AnkiDroid/src/main/java/com/ichi2/anki/AbstractFlashcardViewer.kt
@@ -144,6 +144,7 @@ import com.ichi2.libanki.Sound.getAvTag
 import com.ichi2.libanki.SoundOrVideoTag
 import com.ichi2.libanki.TTSTag
 import com.ichi2.libanki.Utils
+import com.ichi2.libanki.sched.Ease
 import com.ichi2.libanki.undoableOp
 import com.ichi2.themes.Themes
 import com.ichi2.themes.Themes.getResFromAttr
