+++ AnkiDroid/src/main/java/com/ichi2/anki/reviewer/EaseButton.kt
@@ -24,8 +24,8 @@ import android.widget.TextView
 import androidx.annotation.StringRes
 import androidx.core.view.isVisible
 import com.ichi2.anki.AbstractFlashcardViewer
-import com.ichi2.anki.Ease
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
+import com.ichi2.libanki.sched.Ease
 
 /**
  * The UI of an ease button
