+++ AnkiDroid/src/main/java/com/ichi2/libanki/sched/DummyScheduler.kt
@@ -18,10 +18,8 @@
 
 package com.ichi2.libanki.sched
 
-import com.ichi2.anki.Ease
 import com.ichi2.libanki.Card
 import com.ichi2.libanki.Collection
-import java.lang.Exception
 
 class DummyScheduler(
     col: Collection,
