+++ AnkiDroid/src/main/java/com/ichi2/libanki/sched/Scheduler.kt
@@ -41,7 +41,6 @@ import anki.scheduler.SchedulingStates
 import anki.scheduler.UnburyDeckRequest
 import anki.scheduler.cardAnswer
 import anki.scheduler.scheduleCardsAsNewRequest
-import com.ichi2.anki.Ease
 import com.ichi2.anki.common.time.SECONDS_PER_DAY
 import com.ichi2.anki.common.time.TimeManager.time
 import com.ichi2.libanki.Card
