+++ AnkiDroid/src/test/java/com/ichi2/anki/DeckPickerTest.kt
@@ -28,6 +28,7 @@ import com.ichi2.anki.utils.Destination
 import com.ichi2.anki.utils.ext.dismissAllDialogFragments
 import com.ichi2.libanki.DeckId
 import com.ichi2.libanki.Storage
+import com.ichi2.libanki.sched.Ease
 import com.ichi2.testutils.BackendEmulatingOpenConflict
 import com.ichi2.testutils.BackupManagerTestUtilities
 import com.ichi2.testutils.DbUtils
