+++ AnkiDroid/src/test/java/com/ichi2/anki/ReviewerTest.kt
@@ -46,6 +46,7 @@ import com.ichi2.libanki.NotetypeJson
 import com.ichi2.libanki.Notetypes
 import com.ichi2.libanki.QueueType
 import com.ichi2.libanki.exception.ConfirmModSchemaException
+import com.ichi2.libanki.sched.Ease
 import com.ichi2.libanki.undoableOp
 import com.ichi2.testutils.MockTime
 import com.ichi2.testutils.common.Flaky
