+++ AnkiDroid/src/test/java/com/ichi2/libanki/FinderTest.kt
@@ -16,10 +16,10 @@
 package com.ichi2.libanki
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.ichi2.anki.Ease
 import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.libanki.QueueType.Suspended
 import com.ichi2.libanki.exception.ConfirmModSchemaException
+import com.ichi2.libanki.sched.Ease
 import com.ichi2.libanki.sched.Scheduler
 import com.ichi2.testutils.AnkiAssert
 import com.ichi2.testutils.JvmTest
