+++ AnkiDroid/src/test/java/com/ichi2/libanki/SchedulerTest.kt
@@ -17,7 +17,6 @@ package com.ichi2.libanki
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import anki.scheduler.UnburyDeckRequest
-import com.ichi2.anki.Ease
 import com.ichi2.anki.common.time.SECONDS_PER_DAY
 import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.anki.common.time.TimeManager.time
@@ -27,6 +26,7 @@ import com.ichi2.libanki.Consts.LEECH_SUSPEND
 import com.ichi2.libanki.Consts.STARTING_FACTOR
 import com.ichi2.libanki.exception.ConfirmModSchemaException
 import com.ichi2.libanki.sched.Counts
+import com.ichi2.libanki.sched.Ease
 import com.ichi2.testutils.AnkiAssert
 import com.ichi2.testutils.JvmTest
 import com.ichi2.testutils.ext.addNote
